Refactor domain pipelines into scoped packages
