+++ library/pipelines/testitem/__init__.py
@@ -1,0 +1,7 @@
+"""Test item pipeline orchestration helpers."""
+
+from __future__ import annotations
+
+from .enrichment import enrich
+
+__all__ = ["enrich"]
