+++ tests/test_get_target_data_all.py
@@ -18,7 +18,7 @@ import pandas as pd
 import pytest
 from pytest import MonkeyPatch
 
-from library import protein_classification as pc
+from library.pipelines.target import protein_classification as pc
 from library.config import Config
 from library.schemas import TargetsSchema
 from library.schemas.targets import TARGETS_COLUMN_ORDER
